초기 커밋.

+++ Assets/Park/Lobby/Scripts/LobbyPanel.cs
@@ -74,7 +74,7 @@ public class LobbyPanel : MonoBehaviour
         {
 			StatePanel.Instance.AddMessage(room.CustomProperties.ToString());
             RoomEntry entry = Instantiate(roomEntryPrefab, roomContent);
-			entry.Initialize(room.Name, room.PlayerCount, room.MaxPlayers, room, enterPrivateRoomPanel);
+			entry.Initialize(room.Name, room.PlayerCount, (byte)room.MaxPlayers, room, enterPrivateRoomPanel);
             roomEntries.Add(entry);
 		}
     }
